r/role_definition: splitting out the Update and polling until it's settled down
Turns out the Update call is eventually consistent - where a new Role Definition
gets created and then reconciled on the backend. This new Role Definition has a
new CreatedOn and UpdatedOn, so it's possible to determine when this has been
updated - however in testing it can take ~5m for this to settle down.

Whilst it's unfortunate this takes a little longer to settle down, this should
improve the reliability of resources which depend on this.

Unfortunately the Azure SDK doesn't expose these fields in the response, so for
now we're using a wrapped to access them and confirm this has been updated.